Fix #40: non-metaExpr() code evaluated with meta-mode enabled?
Introduces a new metaMode state, "mixed" (in addition to TRUE and
FALSE, which we could consider moving to "meta" and "normal").

This new state is entered when a metaReactive2 (or other -2) is
executing in meta mode, and you're outside the metaExpr portion.
If you're in "mixed" and you invoke another metaReactive/2, then
that invocation will happen in FALSE (normal) mode.

One way to think of this is as a state machine, where the nodes
are NORMAL, META, and MIXED, and the edges are withMetaMode,
metaDispatch (i.e. entering a meta-reactive object), and metaExpr.

(The lines marked with ** are the new and exciting edges)

Normal ==metaDispatch=> Normal
Normal ==metaExpr=> Normal

Meta ==metaDispatch=> Mixed **
Meta ==metaExpr=> Meta

Mixed ==metaDispatch=> Normal **
Mixed ==metaExpr=> Meta

For completeness (because this is the only way out of Normal)
here are the ways withMetaMode() affects the state machine.

[Any] ==withMetaMode(TRUE)=> Meta
[Any] ==withMetaMode(FALSE)=> Normal
[Any] ==withMetaMode("mixed")=> Mixed